feat: split views.py into separate files
