+++ rss/views/__init__.py
@@ -1,0 +1,4 @@
+from .auth import *  # noqa
+from .client import *  # noqa
+from .feed import *  # noqa
+from .torrent import *  # noqa
