+++ rss/views/feed.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+
+class FeedListView(TemplateView):
+    template_name = "rss/feed_list.html"
