+++ rss/views/torrent.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+
+class TorrentListView(TemplateView):
+    template_name = "rss/torrent_list.html"
